added article open toggler, added Article to Board
