+++ src/components/Article.js
@@ -1,3 +1,15 @@
 //single article card
 //contains article title, link, image (optional), and category
 //user can edit or delete article
+
+import React from 'react'
+
+function Article() {
+    return (
+        <div>
+            <button>Add Article</button>
+        </div>
+    )
+}
+
+export default Article
