+++ src/components/Boards.js
@@ -1,2 +1,17 @@
-//list of article boards displayed on dashboard
-//default boards can include work, school, leisure
+import React, { useState } from 'react'
+import Board from './Board'
+import NewBoard from './NewBoard'
+
+function Boards() {
+    const [boards, setBoards] = useState([])
+
+    return (
+        <div>
+            <h2>My Boards</h2>
+            {boards.map(board => <Board board={board}/>)}
+            <NewBoard boards={boards} setBoards={setBoards}/>
+        </div>
+    )
+}
+
+export default Boards
