+++ src/components/Dashboard.js
@@ -1,79 +1,12 @@
-import React, { useState } from 'react'
-
-const initialValues = {
-    name: '',
-    description: '',
-    private: false,  
-}
+import React from 'react'
+import Boards from './Boards'
 
 function Dashboard() {
-    const [isEditing, setIsEditing] = useState(false)
-    const [values, setValues] = useState(initialValues)
-
-    const handleChanges = (e) => {
-        const { name, value, type, checked } = e.target
-        const valueToUse = type === 'checkbox' ? checked : value
-        setValues({ ...values, [name]: valueToUse})
-    }
-
-    const handleSubmit = (e) => {
-        e.preventDefault()
-        setIsEditing(false)
-        setValues(initialValues)
-    }
-
-    const closeForm = (e) => {
-        e.preventDefault()
-        setIsEditing(false)
-        setValues(initialValues)
-    }
 
     return (
         <div>
             <h1>Dashboard</h1>
-            <h2>My Boards</h2>
-            <button onClick={() => {setIsEditing(true)}}>New Board</button>
-            {isEditing ?
-            <form>
-                <label>Name</label>
-                <input 
-                    type='text'
-                    name='name'
-                    value={values.name}
-                    onChange={handleChanges}
-                />
-                <label>Description</label>
-                <input 
-                    type='text'
-                    name='description'
-                    value={values.description}
-                    onChange={handleChanges}
-                />
-                {/* <label>Private</label> */}
-                <label>
-                    Private
-                <input 
-                    type='checkbox'
-                    name='private'
-                    checked={values.private}
-                    onChange={handleChanges}
-                />
-                </label>
-                {/* <label>
-                    No
-                <input 
-                    type='radio'
-                    name='private'
-                    value='false'
-                    checked={values.private}
-                    onChange={handleChanges}
-                />
-                </label> */}
-                <button onClick={handleSubmit}>Save</button>
-                <button onClick={closeForm}>Close</button>
-            </form>
-            : null
-            }
+            <Boards />
         </div>
     )
 }
